Extracted BearerTokenProvider out and using it in both DevLoginExample and LoginExample
