+++ src/main/java/com/salesforce/emp/connector/example/LoginExample.java
@@ -8,12 +8,14 @@ package com.salesforce.emp.connector.example;
 
 import static com.salesforce.emp.connector.LoginHelper.login;
 
+import java.net.URL;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Consumer;
 
 import com.salesforce.emp.connector.BayeuxParameters;
 import com.salesforce.emp.connector.EmpConnector;
+import com.salesforce.emp.connector.LoginHelper;
 import com.salesforce.emp.connector.TopicSubscription;
 
 /**
@@ -33,18 +35,23 @@ public class LoginExample {
             replayFrom = Long.parseLong(argv[3]);
         }
 
-        BayeuxParameters params;
-        try {
-            params = login(argv[0], argv[1]);
-        } catch (Exception e) {
-            e.printStackTrace(System.err);
-            System.exit(1);
-            throw e;
-        }
+        BearerTokenProvider tokenProvider = new BearerTokenProvider(() -> {
+            try {
+                return login(argv[0], argv[1]);
+            } catch (Exception e) {
+                e.printStackTrace(System.err);
+                System.exit(1);
+                throw new RuntimeException(e);
+            }
+        });
+
+        BayeuxParameters params = tokenProvider.login();
 
         Consumer<Map<String, Object>> consumer = event -> System.out.println(String.format("Received:\n%s", event));
 
         EmpConnector connector = new EmpConnector(params);
+
+        connector.setBearerTokenProvider(tokenProvider);
 
         connector.start().get(5, TimeUnit.SECONDS);
